@@ -1,37 +1,49 @@
+//Controls routing in application
+
 import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { Routes, Route } from "react-router-dom";
 import Login from "./Login";
-import { Provider } from "react-redux";
+import { Provider } from "react-redux"; // TODO
 import Home from "./Home";
 import ClassTemplate from "./ClassTemplate";
+import ClassTest from "./ClassTest"
 import store from "./store";
 import SignUp from "./SignUp";
 import Profile from "./Profile";
-import AboutUs from "./AboutUs";
-import Logout from "./Logout";
-import Registration from "./Registration";
+import ProfileTest from "./ProfileTest";
 import ResetPassword from "./ResetPassword";
+import Logout from "./Logout";
+import AboutPage from "./AboutPage";
+import Register from "./Register";
 import ResetPasswordToken from "./ResetPasswordToken";
 
-const AppRouter = () => {
 
+const AppRouter = () => {
   return (
     <Provider store={store}>
-      <BrowserRouter>
         <Routes>
           <Route path="/" element={<Home />} />
-          <Route path="/classes" element={<ClassTemplate />} />
+          <Route path="/home" element={<Home />} />
+          <Route path='/logout' element={<Logout />} />
+          <Route path='/aboutus' element={<AboutPage />} />
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<SignUp />} />
+          <Route path="/classes" element={<ClassTemplate />} />
+          <Route path="/classes_test" element={<ClassTest />} />
           <Route path="/profile" element={<Profile />} />
-          <Route path='/aboutus' element={<AboutUs />} />
-          <Route path='/logout' element={<Logout />} />
-          <Route path='/registration' element={<Registration />} />
-          <Route path='/reset_password' element={<ResetPassword />} />
+          <Route path="/profile_test" element={<ProfileTest />} />
+          <Route path="/reset_password" element={<ResetPassword />} />
           <Route path='/reset_password/:token' element={<ResetPasswordToken />} />
+          <Route path='/register' element={<Register />} />
+
+          {/* 
+          
+          
+          
+            */}
         </Routes>
-      </BrowserRouter>
     </Provider>
   );
 };
+
 export default AppRouter;
